backend/requirements.txt
pydantic
fastapi
pillow
roboflow
uvicorn
python-dotenv
python-multipart
fpdf2
google-generativeai
